Update status logics

+++ app/Services/Service.php
@@ -5,14 +5,12 @@ namespace App\Services;
 use App\Debugger;
 use App\Helpers\Helper;
 use Closure;
-use DateInterval;
-use DateTimeInterface;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Str;
 
-class Service
+abstract class Service
 {
     use Debugger;
 
@@ -20,13 +18,45 @@ class Service
 
     protected string $modelName;
 
-    protected callable|DateInterval|DateTimeInterface|int|null $cacheTTL;
+    protected int $cacheTTL = 60;
 
-    protected function __construct(?Model $model = null, callable|DateInterval|DateTimeInterface|int|null $cacheTTL = 60)
+    protected function __construct(?Model $model = null, int $cacheTTL = 60)
     {
         $this->model = $model;
-        $this->modelName = Str::lower(class_basename($model ?? ''));
+        $this->modelName = Str::lower(class_basename($model ?? 'record'));
         $this->cacheTTL = $cacheTTL;
+    }
+
+    private function generateCacheKey(string $key, array $filters = []): string
+    {
+        try {
+            $filtered = Helper::array_filter($filters);
+            $queryString = $filtered ? '.'.http_build_query($filtered) : '';
+
+            return "{$this->modelName}.{$key}{$queryString}";
+        } catch (\Throwable $th) {
+            $this->debug('error', 'Failed to generate cache key.', $th);
+            throw $th;
+        }
+    }
+
+    protected function cachedQuery(Closure $callback, string $key, array $filters, int $ttl = 60): mixed
+    {
+        try {
+            return Cache::remember(
+                $this->generateCacheKey($key, $filters),
+                $ttl ?? $this->cacheTTL,
+                fn () => $callback($this->model)
+            );
+        } catch (\Throwable $th) {
+            $this->debug('error', $th->getMessage(), $th);
+            throw $th;
+        }
+    }
+
+    protected function model(array $attributes = []): ?Model
+    {
+        return $this->model->fill($attributes);
     }
 
     protected function collection(array $items = []): Collection
@@ -43,7 +73,7 @@ class Service
 
             return $this->model->all();
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to get all {$this->modelName}(s) data.", $th);
+            $this->debug('error', "Failed to get all {$this->modelName}(s).", $th);
             throw $th;
         }
     }
@@ -57,7 +87,7 @@ class Service
 
             return $this->model->where($where)->get();
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to get {$this->modelName}(s) data.", $th);
+            $this->debug('error', "Failed to get {$this->modelName}(s).", $th);
             throw $th;
         }
     }
@@ -76,16 +106,16 @@ class Service
         }
     }
 
-    protected function firstOrInit(array $with = [], array $attributes = []): Model|Collection
+    protected function firstOrInit(array $attributes = [], array $with = []): Model|Collection
     {
         try {
             if (! isset($this->model)) {
                 return $this->collection($attributes);
             }
 
-            return $this->first($with) ?? new $this->model($attributes);
+            return $this->first($with) ?? $this->model($attributes);
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to get the first or init {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to get the first or init {$this->modelName}.", $th);
             throw $th;
         }
     }
@@ -99,7 +129,7 @@ class Service
 
             return $this->model->firstOrCreate($attributes, $values);
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to get the first or create new {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to get the first or create new {$this->modelName}.", $th);
             throw $th;
         }
     }
@@ -113,27 +143,75 @@ class Service
 
             return $this->model->where($where)->first();
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to get with conditions for the first {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to get with conditions for the first {$this->modelName}.", $th);
             throw $th;
         }
     }
 
-    protected function create(array $data): ?Model
+    protected function find(int|string $id, array $with = []): ?Model
     {
         try {
-            return $this->model->create($data);
+            if (! empty($with)) {
+                return $this->model->with($with)->find($id);
+            }
+
+            return $this->model->find($id);
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to create a new {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to find {$this->modelName} with id {$id}.", $th);
             throw $th;
         }
     }
 
-    protected function update(array $attributes, array $options = []): bool
+    protected function findOrFail(int|string $id, array $with = []): ?Model
     {
         try {
-            return $this->model->update($attributes, $options);
+            if (! empty($with)) {
+                return $this->model->with($with)->findOrFail($id);
+            }
+
+            return $this->model->findOrFail($id);
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to update {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to find {$this->modelName} with id {$id}.", $th);
+            throw $th;
+        }
+    }
+
+    protected function exists(array $where): bool
+    {
+        try {
+            return $this->model->where($where)->exists();
+        } catch (\Throwable $th) {
+            $this->debug('error', "Failed to check if {$this->modelName} exists.", $th);
+            throw $th;
+        }
+    }
+
+    protected function create(array $attributes): ?Model
+    {
+        try {
+            return $this->model->create($attributes);
+        } catch (\Throwable $th) {
+            $this->debug('error', "Failed to create a new {$this->modelName}.", $th);
+            throw $th;
+        }
+    }
+
+    protected function update(array $where, array $attributes, array $options = []): bool
+    {
+        try {
+            return $this->model->where($where)->update($attributes, $options);
+        } catch (\Throwable $th) {
+            $this->debug('error', "Failed to update {$this->modelName}.", $th);
+            throw $th;
+        }
+    }
+
+    protected function updateFirst(array $where, array $attributes, array $options = []): bool
+    {
+        try {
+            return $this->model->where($where)->first()->update($attributes, $options);
+        } catch (\Throwable $th) {
+            $this->debug('error', "Failed to update the first {$this->modelName}.", $th);
             throw $th;
         }
     }
@@ -143,44 +221,31 @@ class Service
         try {
             return $this->model->updateOrCreate($attributes, $values);
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to update or create a new {$this->modelName} data.", $th);
+            $this->debug('error', "Failed to update or create a new {$this->modelName}.", $th);
             throw $th;
         }
     }
 
-    protected function delete(array $where): bool
+    protected function delete(int|string|array $where): bool
     {
         try {
-            return $this->model->where($where)->delete();
+            if (is_array($where)) {
+                return $this->model->where($where)->delete();
+            }
+
+            return $this->model->find($where)->delete();
         } catch (\Throwable $th) {
-            $this->debug('error', "Failed to delete {$this->modelName} data.");
+            $this->debug('error', "Failed to delete {$this->modelName}.");
             throw $th;
         }
     }
 
-    private function generateCacheKey(string $key, array $filters = []): string
+    protected function destroy(int|string $ids): bool
     {
         try {
-            $filtered = Helper::array_filter($filters);
-            $queryString = $filtered ? '.'.http_build_query($filtered) : '';
-
-            return "{$this->modelName}.{$key}{$queryString}";
+            return $this->model->destroy($ids);
         } catch (\Throwable $th) {
-            $this->debug('error', 'Failed to generate cache key.', $th);
-            throw $th;
-        }
-    }
-
-    protected function cachedQuery(Closure $callback, string $key, array $filters, callable|DateInterval|DateTimeInterface|int|null $ttl = 60): mixed
-    {
-        try {
-            return Cache::remember(
-                $this->generateCacheKey($key, $filters),
-                $ttl ?? $this->cacheTTL,
-                fn () => $callback($this->model)
-            );
-        } catch (\Throwable $th) {
-            $this->debug('error', $th->getMessage(), $th);
+            $this->debug('error', "Failed to destroy {$this->modelName}(s).");
             throw $th;
         }
     }
